Refactor App.js to use Home component
